--- app/(drawer)/(tabs)/index.tsx
@@ -1,9 +1,0 @@
-import { View, Text } from 'react-native';
-
-export default function Home() {
-  return (
-    <View className='p-4'>
-      <Text className='text-gray-400 text-2xl'>Hello</Text>
-    </View>
-  );
-}
